basleracA130075gmApp/Db/, iocBoot/iocBasleracA130075gm/: Add high-level aliases in low-level
Add "HighLevelAliases.db" file containing aliases for the low-level PVs.

This simplifies future expansions of the camera code, maintaining related PV names in the same place.

In order to allow enabling and disabling of high level aliases, an extra option can be provided to the startup script.
if the -A option is specified, the following parameter is the prefix used by high-level aliases. When this option
is not specified, the high-level aliases are not created.

This feature is provided by the "stExtarAliases.cmd" file, along with changes in the "parseCMDOpts.sh", and "runBasleracA130075gm.sh" scripts.