basleracA130075gmApp/Db/BaslerSiriusStandard.db: Add auto reconnection
Add an auto reconnection attempt every 10 seconds after the connection is lost.
When the connection is restored, the IOC PV values are downloaded to the camera again.
In order to avoid problems with auto downloading PV values at startup, the PVs now have default values that are consistent with the camera default values, so that no problematic configuration is applied when there are no autosaved values.

Extra feature:

Camera statistics PVs now are divided in two categories: NDStats and DimFei.