{basleracA130075gmApp/Db/, iocBoot/, op/opi/}: Add profile statistics based on a ROI
Two additional plugins are used: an extra stats plugin and a ROI plugin. The ROI is used for selecting the region for which statistics are going to be produced. The second stats plugin uses the ROI output image to calculate the image profile in the X and Y axes.

Some of the plugins PVs, which are expected to be used more often, are aliased to names that comply with the Sirius naming system.

These PVs include:
  - enable/disable of the ROI plugin (so that the new feature can be disabled).
  - ROI width, height, and offset in the X and Y axes.
  - stats2 profile cursors position.
  - profile along the X cursor.
  - profile along the Y cursor.
  - average profile of the X axis.
  - average profile of the Y axis.