Change parameters of centroid position unit conversion
Now the user sets: X offset, Y offset, X scale factor, and Y scale factor.

X offset and Y offset should be equal to the offset in pixels in the X and Y axis of the reference centroid position, considering the coordinates of the image (left-topmost pixel is at (1,1)).

X and Y scale factors should be equal to a factor that multiplied by a position in pixels yields the position in the desired unit. Using a negative value inverts the orientation of the associated axis, allowing reference frames different from the camera's (Y increasing downwards and X increasing to the right) to be used. For example, in order to have a Y axis that increases from the bottom to the top of the image, the y scale factor should be negative and have an absolute value equal to the desired conversion factor.