Add a ROI plugin to control main image output
Add a 'soft ROI' layer to allow ROI control independently
of the Transform plugin settings.

An alternative would be to create soft ROI records which would
control the camera ROI, but doing it differently for each Transform
plugin configuration (rotation and mirror) so that it would keep
the X and Y axis the same from the operator point of view.

The alternative above seemed not the best at the moment, since the use
of the ROI plugin makes the code clearer and less coupled for future
maintenance and changes.